[Skia] Add a pixel test: same color knockout paths
I wanted to make sure that in the new code where I drop the branch for
fill_color == stroke_color (In CFX_SkiaDeviceDriver::DrawPath()) that it
doesn't accidentally render an anti-aliasing border between the stroke
and fill (from the inside edge of the stroke getting AA). It should just
be a single solid color with AA around the outside edge.

bug_1286.in already incidentally covered this code path but I wanted to
make a more explicit test. This test case does it in a different way.

For some reason gdi_skia renders absolutely nothing for this PDF!
Clearly this is wrong and the second attached bug is about this.

Bug: 334774307, 460743387
Change-Id: I8ac13daaf55f118deefbc5883d277688ca5ba135
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/138270
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: April Kallmeyer <ask@chromium.org>