Make GetRec() private with ScopedFontTransform as a friend class
This CL makes GetRec() private, with the addition of ScopedFontTransform
as a friend class. This is done to avoid exposing struct FT_FaceRec_ and
after previous consolidation of other public callers and movement of
some logic in these callers to cfx_face.cpp. Allowing GetRec() to be
private.

Bug: 460068801
Change-Id: I6e16c65f1452e803f18f93aa73e7bbe8af867bc4
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/138570
Commit-Queue: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>